This issue manifested when using jSS7 to parse assorted SS7 PCAP trace files.
Some background;

TU-T Rec. Q.713 (07/96) Signalling Connection Control Part formats and
codes, section 3.4.2.3.4 Global title indicator = 0100, shows an example
encoding of an Nature of address indicator value that is stored in bits
1-7 of a single byte.

Bit 8 is shown as 0.

3.4.2.3.1 Global title indicator = 0001 of the same document contains a
detailed bitmask diagram of Nature of Address Indicator (NAI), again
showing bits 1 to 7 of a single byte.

In this case, bit 8  contains the odd/even indicator.

It seems clear to me that only bits 1-7 should be evaluated when
calculating the NAI.

However, the decode() method of the GlobalTitle0100Impl() class uses a
bitmask of 0xff - I think this should be 0x7f.